+++ client/src/main/java/it/sosinski/chat/manager/ManagerService.java
@@ -2,5 +2,5 @@ package it.sosinski.chat.manager;
 
 public interface ManagerService {
 
-    void process(Integer channelId, String text, String name);
+    void process(Long channelId, String text, String name);
 }
