+++ client/src/main/java/it/sosinski/chat/manager/RestService.java
@@ -2,6 +2,6 @@ package it.sosinski.chat.manager;
 
 public interface RestService {
 
-    void process(Integer channelId, String text, String name);
+    void process(Long channelId, String text, String name);
 
 }
